(feat): workout the wasm runner imterface and implementnation
