+++ engine-tests/Cargo.toml
@@ -50,6 +50,7 @@ sha3.workspace = true
 tempfile.workspace = true
 tokio.workspace = true
 walrus.workspace = true
+memoffset.workspace = true
 
 [lints]
 workspace = true
