+++ engine-tests/src/utils/mod.rs
@@ -43,6 +43,7 @@ const CALLER_ACCOUNT_ID: &str = "some-account.near";
 
 pub mod mocked_external;
 pub mod one_inch;
+pub mod runner;
 pub mod rust;
 pub mod solidity;
 pub mod standalone;
